Move DatasetArtifact to _private/dataset.py — mirrors internal repo convention
workflow/variables/_private/dataset.py now holds DatasetArtifact, matching
the internal repo where DatasetVariable lives in _private/dataset.py.

Users import DatasetArtifact from michelangelo.workflow.variables (__init__),
not from the _private path directly. types.py re-export updated accordingly.
