+++ python/michelangelo/workflow/variables/__init__.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 import contextlib
 
 # flake8: noqa:F401
-from michelangelo.workflow.variables.dataset import DatasetArtifact
+from michelangelo.workflow.variables._private.dataset import DatasetArtifact
 from michelangelo.workflow.variables.metadata import ModelMetadata
 from michelangelo.workflow.variables.types import (
     AssembledModel,
